requirements.txt
streamlit
pandas
scikit-learn
xgboost
imblearn
joblib
imbalanced-learn
numpy
matplotlib
pillow
